Split coverage.py into several files
