+++ src/gcovr/data_model/__init__.py
@@ -1,0 +1,18 @@
+# -*- coding:utf-8 -*-
+
+#  ************************** Copyrights and license ***************************
+#
+# This file is part of gcovr 8.3+main, a parsing and reporting tool for gcov.
+# https://gcovr.com/en/main
+#
+# _____________________________________________________________________________
+#
+# Copyright (c) 2013-2025 the gcovr authors
+# Copyright (c) 2013 Sandia Corporation.
+# Under the terms of Contract DE-AC04-94AL85000 with Sandia Corporation,
+# the U.S. Government retains certain rights in this software.
+#
+# This software is distributed under the 3-clause BSD License.
+# For more information, see the README.rst file.
+#
+# ****************************************************************************
