+++ src/gcovr/data_model/coverage.py
@@ -36,104 +36,11 @@ report aggregated metrics/percentages.
 
 from __future__ import annotations
 import logging
-import os
-import re
-from typing import (
-    ItemsView,
-    Iterator,
-    Iterable,
-    Optional,
-    TypeVar,
-    Union,
-    Literal,
-    ValuesView,
-)
-from dataclasses import dataclass
+from typing import Optional, Union
 
-from .utils import commonpath, force_unix_separator
+from .stats import CoverageStat, DecisionCoverageStat, SummarizedStats
 
 LOGGER = logging.getLogger("gcovr")
-
-_T = TypeVar("_T")
-
-
-def sort_coverage(
-    covdata: Union[
-        dict[str, FileCoverage],
-        dict[str, Union[FileCoverage, CoverageContainerDirectory]],
-    ],
-    sort_key: Literal["filename", "uncovered-number", "uncovered-percent"],
-    sort_reverse: bool,
-    by_metric: Literal["line", "branch", "decision"],
-    filename_uses_relative_pathname: bool = False,
-) -> list[str]:
-    """Sort a coverage dict.
-
-    covdata (dict): the coverage dictionary
-    sort_key ("filename", "uncovered-number", "uncovered-percent"): the values to sort by
-    sort_reverse (bool): reverse order if True
-    by_metric ("line", "branch", "decision"): select the metric to sort
-    filename_uses_relative_pathname (bool): for html, we break down a pathname to the
-        relative path, but not for other formats.
-
-    returns: the sorted keys
-    """
-
-    basedir = commonpath(list(covdata.keys()))
-
-    def key_filename(key: str) -> list[Union[int, str]]:
-        def convert_to_int_if_possible(text: str) -> Union[int, str]:
-            return int(text) if text.isdigit() else text
-
-        key = (
-            force_unix_separator(
-                os.path.relpath(os.path.realpath(key), os.path.realpath(basedir))
-            )
-            if filename_uses_relative_pathname
-            else key
-        ).casefold()
-
-        return [convert_to_int_if_possible(part) for part in re.split(r"([0-9]+)", key)]
-
-    def coverage_stat(key: str) -> CoverageStat:
-        cov = covdata[key]
-        if by_metric == "branch":
-            return cov.branch_coverage()
-        if by_metric == "decision":
-            return cov.decision_coverage().to_coverage_stat
-        return cov.line_coverage()
-
-    def key_num_uncovered(key: str) -> int:
-        stat = coverage_stat(key)
-        uncovered = stat.total - stat.covered
-        return uncovered
-
-    def key_percent_uncovered(key: str) -> float:
-        stat = coverage_stat(key)
-        covered = stat.covered
-        total = stat.total
-
-        # No branches are always put directly after (or before when reversed)
-        # files with 100% coverage (by assigning such files 110% coverage)
-        return covered / total if total > 0 else 1.1
-
-    if sort_key == "uncovered-number":
-        # First sort filename alphabetical and then by the requested key
-        return sorted(
-            sorted(covdata, key=key_filename),
-            key=key_num_uncovered,
-            reverse=sort_reverse,
-        )
-    if sort_key == "uncovered-percent":
-        # First sort filename alphabetical and then by the requested key
-        return sorted(
-            sorted(covdata, key=key_filename),
-            key=key_percent_uncovered,
-            reverse=sort_reverse,
-        )
-
-    # By default, we sort by filename alphabetically
-    return sorted(covdata, key=key_filename, reverse=sort_reverse)
 
 
 class BranchCoverage:
@@ -684,338 +591,3 @@ class FileCoverage:
                             covered += 1
 
         return CoverageStat(covered, total)
-
-
-class CoverageContainer:
-    """Coverage container holding all the coverage data."""
-
-    def __init__(self) -> None:
-        self.data = dict[str, FileCoverage]()
-        self.directories = list[CoverageContainerDirectory]()
-
-    def __getitem__(self, key: str) -> FileCoverage:
-        return self.data[key]
-
-    def __len__(self) -> int:
-        return len(self.data)
-
-    def __contains__(self, key: str) -> bool:
-        return key in self.data
-
-    def __iter__(self) -> Iterator[str]:
-        return iter(self.data)
-
-    def values(self) -> ValuesView[FileCoverage]:
-        """Get the file coverage data objects."""
-        return self.data.values()
-
-    def items(self) -> ItemsView[str, FileCoverage]:
-        """Get the file coverage data items."""
-        return self.data.items()
-
-    @property
-    def stats(self) -> SummarizedStats:
-        """Create a coverage statistic from a coverage data object."""
-        stats = SummarizedStats.new_empty()
-        for filecov in self.values():
-            stats += filecov.stats
-        return stats
-
-    def sort_coverage(
-        self,
-        sort_key: Literal["filename", "uncovered-number", "uncovered-percent"],
-        sort_reverse: bool,
-        by_metric: Literal["line", "branch", "decision"],
-        filename_uses_relative_pathname: bool = False,
-    ) -> list[str]:
-        """Sort the coverage data"""
-        return sort_coverage(
-            self.data,
-            sort_key,
-            sort_reverse,
-            by_metric,
-            filename_uses_relative_pathname,
-        )
-
-    @staticmethod
-    def _get_dirname(filename: str) -> Optional[str]:
-        """Get the directory name with a trailing path separator.
-
-        >>> import os
-        >>> CoverageContainer._get_dirname("bar/foobar.cpp".replace("/", os.sep)).replace(os.sep, "/")
-        'bar/'
-        >>> CoverageContainer._get_dirname("/foo/bar/A/B.cpp".replace("/", os.sep)).replace(os.sep, "/")
-        '/foo/bar/A/'
-        >>> CoverageContainer._get_dirname(os.sep) is None
-        True
-        """
-        if filename == os.sep:
-            return None
-        return str(os.path.dirname(filename.rstrip(os.sep))) + os.sep
-
-    def populate_directories(
-        self, sorted_keys: Iterable[str], root_filter: re.Pattern[str]
-    ) -> None:
-        r"""Populate the list of directories and add accumulated stats.
-
-        This function will accumulate statistics such that every directory
-        above it will know the statistics associated with all files deep within a
-        directory structure.
-
-        Args:
-            sorted_keys: The sorted keys for covdata
-            root_filter: Information about the filter used with the root directory
-        """
-
-        # Get the directory coverage
-        subdirs = dict[str, CoverageContainerDirectory]()
-        for key in sorted_keys:
-            filecov = self[key]
-            dircov: Optional[CoverageContainerDirectory] = None
-            dirname: Optional[str] = (
-                os.path.dirname(filecov.filename)
-                .replace("\\", os.sep)
-                .replace("/", os.sep)
-                .rstrip(os.sep)
-            ) + os.sep
-            while dirname is not None and root_filter.search(dirname + os.sep):
-                if dirname not in subdirs:
-                    subdirs[dirname] = CoverageContainerDirectory(dirname)
-                if dircov is None:
-                    subdirs[dirname][filecov.filename] = filecov
-                else:
-                    subdirs[dirname].data[dircov.filename] = dircov
-                    subdirs[dircov.filename].parent_dirname = dirname
-                subdirs[dirname].stats += filecov.stats
-                dircov = subdirs[dirname]
-                dirname = CoverageContainer._get_dirname(dirname)
-
-        # Replace directories where only one sub container is available
-        # with the content this sub container
-        LOGGER.debug(
-            "Replace directories with only one sub element with the content of this."
-        )
-        subdirs_to_remove = set()
-        for dirname, covdata_dir in subdirs.items():
-            # There is exact one element, replace current element with referenced element
-            if len(covdata_dir) == 1:
-                # Get the orphan item
-                orphan_key, orphan_value = next(iter(covdata_dir.items()))
-                # The only child is a File object
-                if isinstance(orphan_value, FileCoverage):
-                    # Replace the reference to ourself with our content
-                    if covdata_dir.parent_dirname is not None:
-                        LOGGER.debug(
-                            f"Move {orphan_key} to {covdata_dir.parent_dirname}."
-                        )
-                        parent_covdata_dir = subdirs[covdata_dir.parent_dirname]
-                        parent_covdata_dir[orphan_key] = orphan_value
-                        del parent_covdata_dir[dirname]
-                        subdirs_to_remove.add(dirname)
-                else:
-                    LOGGER.debug(
-                        f"Move content of {orphan_value.dirname} to {dirname}."
-                    )
-                    # Replace the children with the orphan ones
-                    covdata_dir.data = orphan_value.data
-                    # Change the parent key of each new child element
-                    for new_child_value in covdata_dir.values():
-                        if isinstance(new_child_value, CoverageContainerDirectory):
-                            new_child_value.parent_dirname = dirname
-                    # Mark the key for removal.
-                    subdirs_to_remove.add(orphan_key)
-
-        for dirname in subdirs_to_remove:
-            del subdirs[dirname]
-
-        self.directories = list(subdirs.values())
-
-
-class CoverageContainerDirectory:
-    """Represent coverage information about a directory."""
-
-    __slots__ = "dirname", "parent_dirname", "data", "stats"
-
-    def __init__(self, dirname: str) -> None:
-        super().__init__()
-        self.dirname: str = dirname
-        self.parent_dirname: Optional[str] = None
-        self.data = dict[str, Union[FileCoverage, CoverageContainerDirectory]]()
-        self.stats: SummarizedStats = SummarizedStats.new_empty()
-
-    def __setitem__(
-        self, key: str, item: Union[FileCoverage, CoverageContainerDirectory]
-    ) -> None:
-        self.data[key] = item
-
-    def __getitem__(self, key: str) -> Union[FileCoverage, CoverageContainerDirectory]:
-        return self.data[key]
-
-    def __delitem__(self, key: str) -> None:
-        del self.data[key]
-
-    def __len__(self) -> int:
-        return len(self.data)
-
-    def values(self) -> ValuesView[Union[FileCoverage, CoverageContainerDirectory]]:
-        """Get the file coverage data objects."""
-        return self.data.values()
-
-    def items(self) -> ItemsView[str, Union[FileCoverage, CoverageContainerDirectory]]:
-        """Get the file coverage data items."""
-        return self.data.items()
-
-    @property
-    def filename(self) -> str:
-        """Helpful function for when we use this DirectoryCoverage in a union with FileCoverage"""
-        return self.dirname
-
-    def sort_coverage(
-        self,
-        sort_key: Literal["filename", "uncovered-number", "uncovered-percent"],
-        sort_reverse: bool,
-        by_metric: Literal["line", "branch", "decision"],
-        filename_uses_relative_pathname: bool = False,
-    ) -> list[str]:
-        """Sort the coverage data"""
-        return sort_coverage(
-            self.data,
-            sort_key,
-            sort_reverse,
-            by_metric,
-            filename_uses_relative_pathname,
-        )
-
-    def line_coverage(self) -> CoverageStat:
-        """A simple wrapper function necessary for sort_coverage()."""
-        return self.stats.line
-
-    def branch_coverage(self) -> CoverageStat:
-        """A simple wrapper function necessary for sort_coverage()."""
-        return self.stats.branch
-
-    def decision_coverage(self) -> DecisionCoverageStat:
-        """A simple wrapper function necessary for sort_coverage()."""
-        return self.stats.decision
-
-
-@dataclass
-class SummarizedStats:
-    """Data class for the summarized coverage statistics."""
-
-    line: CoverageStat
-    branch: CoverageStat
-    condition: CoverageStat
-    decision: DecisionCoverageStat
-    function: CoverageStat
-    call: CoverageStat
-
-    @staticmethod
-    def new_empty() -> SummarizedStats:
-        """Create a empty coverage statistic."""
-        return SummarizedStats(
-            line=CoverageStat.new_empty(),
-            branch=CoverageStat.new_empty(),
-            condition=CoverageStat.new_empty(),
-            decision=DecisionCoverageStat.new_empty(),
-            function=CoverageStat.new_empty(),
-            call=CoverageStat.new_empty(),
-        )
-
-    def __iadd__(self, other: SummarizedStats) -> SummarizedStats:
-        self.line += other.line
-        self.branch += other.branch
-        self.condition += other.condition
-        self.decision += other.decision
-        self.function += other.function
-        self.call += other.call
-        return self
-
-
-@dataclass
-class CoverageStat:
-    """A single coverage metric, e.g. the line coverage percentage of a file."""
-
-    covered: int
-    """How many elements were covered."""
-
-    total: int
-    """How many elements there were in total."""
-
-    @staticmethod
-    def new_empty() -> CoverageStat:
-        """Create a empty coverage statistic."""
-        return CoverageStat(0, 0)
-
-    @property
-    def percent(self) -> Optional[float]:
-        """Percentage of covered elements, equivalent to ``self.percent_or(None)``"""
-        return self.percent_or(None)
-
-    def percent_or(self, default: _T) -> Union[float, _T]:
-        """Percentage of covered elements.
-
-        Coverage is truncated to one decimal:
-        >>> CoverageStat(1234, 10000).percent_or("default")
-        12.3
-
-        Coverage is capped at 99.9% unless everything is covered:
-        >>> CoverageStat(9999, 10000).percent_or("default")
-        99.9
-        >>> CoverageStat(10000, 10000).percent_or("default")
-        100.0
-
-        If there are no elements, percentage is NaN and the default will be returned:
-        >>> CoverageStat(0, 0).percent_or("default")
-        'default'
-        """
-        if not self.total:
-            return default
-
-        # Return 100% only if covered == total.
-        if self.covered == self.total:
-            return 100.0
-
-        # There is at least one uncovered item.
-        # Round to 1 decimal and clamp to max 99.9%.
-        ratio = self.covered / self.total
-        return min(99.9, round(ratio * 100.0, 1))
-
-    def __iadd__(self, other: CoverageStat) -> CoverageStat:
-        self.covered += other.covered
-        self.total += other.total
-        return self
-
-
-@dataclass
-class DecisionCoverageStat:
-    """A CoverageStat for decision coverage (accounts for Uncheckable cases)."""
-
-    covered: int
-    uncheckable: int
-    total: int
-
-    @classmethod
-    def new_empty(cls) -> DecisionCoverageStat:
-        """Create a empty decision coverage statistic."""
-        return cls(0, 0, 0)
-
-    @property
-    def to_coverage_stat(self) -> CoverageStat:
-        """Convert a decision coverage statistic to a coverage statistic."""
-        return CoverageStat(covered=self.covered, total=self.total)
-
-    @property
-    def percent(self) -> Optional[float]:
-        """Return the percent value of the coverage."""
-        return self.to_coverage_stat.percent
-
-    def percent_or(self, default: _T) -> Union[float, _T]:
-        """Return the percent value of the coverage or the given default if no coverage is present."""
-        return self.to_coverage_stat.percent_or(default)
-
-    def __iadd__(self, other: DecisionCoverageStat) -> DecisionCoverageStat:
-        self.covered += other.covered
-        self.uncheckable += other.uncheckable
-        self.total += other.total
-        return self
