+++ src/gcovr/formats/clover/__init__.py
@@ -22,7 +22,7 @@ from typing import Union
 from ...options import GcovrConfigOption, OutputOrDefault
 from ...formats.base import BaseHandler
 
-from ...coverage import CoverageContainer
+from ...data_model.container import CoverageContainer
 
 
 class CloverHandler(BaseHandler):
