+++ src/gcovr/formats/clover/write.py
@@ -23,14 +23,14 @@ from dataclasses import dataclass
 import logging
 from lxml import etree  # nosec # We only write XML files
 
+from ...data_model.container import CoverageContainer
+from ...data_model.coverage import LineCoverage
 from ...options import Options
-
 from ...utils import (
     get_md5_hexdigest,
     open_binary_for_writing,
     presentable_filename,
 )
-from ...coverage import CoverageContainer, LineCoverage
 
 LOGGER = logging.getLogger("gcovr")
