+++ src/gcovr/formats/gcov/parser/json.py
@@ -42,7 +42,7 @@ from typing import (
 
 from gcovr.utils import get_md5_hexdigest
 
-from ....coverage import (
+from ....data_model.coverage import (
     BranchCoverage,
     ConditionCoverage,
     FileCoverage,
