+++ src/gcovr/formats/lcov/write.py
@@ -26,10 +26,9 @@ and generated by https://github.com/linux-test-project/lcov/blob/07a1127c2b4390a
 
 # cspell:ignore FNDA BRDA
 
+from ...data_model.container import CoverageContainer
 from ...options import Options
-
 from ...utils import force_unix_separator, get_md5_hexdigest, open_text_for_writing
-from ...coverage import CoverageContainer
 
 
 def write_report(
