+++ src/gcovr/formats/sonarqube/__init__.py
@@ -19,7 +19,7 @@
 
 from typing import Union
 
-from ...coverage import CoverageContainer
+from ...data_model.container import CoverageContainer
 from ...formats.base import BaseHandler
 from ...options import GcovrConfigOption, OutputOrDefault
