+++ src/gcovr/formats/txt/__init__.py
@@ -20,7 +20,7 @@
 import logging
 from typing import Union
 
-from ...coverage import CoverageContainer
+from ...data_model.container import CoverageContainer
 from ...formats.base import BaseHandler
 from ...options import (
     GcovrConfigOption,
